Remove provider specific classes for now
